d128

+++ cas/src/terminal.rs
@@ -1,6 +1,6 @@
 #[derive(PartialEq, Debug, Clone)]
 pub enum Literal {
     Int(i128),
-    Float(f64),
+    Decimal(f64),
     Symbol(String)
 }
